neat code
