+++ aderyn/src/main.rs
@@ -1,25 +1,10 @@
-#![allow(clippy::borrowed_box)]
-
-use std::path::PathBuf;
-use std::sync::Arc;
-use std::time::Duration;
-
 use aderyn::{
-    aderyn_is_currently_running_newest_version, print_all_detectors_view, print_detail_view,
+    aderyn_is_currently_running_newest_version, lsp::spin_up_language_server,
+    print_all_detectors_view, print_detail_view,
 };
-use log::{info, warn, LevelFilter};
-use notify_debouncer_full::notify::RecommendedWatcher;
-use simple_logging::log_to_file;
-use tokio::sync::Mutex;
-use tower_lsp::jsonrpc::Result;
-use tower_lsp::lsp_types::*;
-use tower_lsp::{Client, LanguageServer, LspService, Server};
-
 use aderyn_driver::driver::{self, Args};
 
 use clap::{ArgGroup, Parser, Subcommand};
-use notify_debouncer_full::notify::{Config, RecursiveMode, Watcher};
-
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 #[command(group(ArgGroup::new("stdout_dependent").requires("stdout")))]
@@ -152,197 +137,4 @@ fn main() {
             }
         }
     }
-}
-
-#[derive(Debug)]
-struct Backend {
-    client: Arc<Mutex<Client>>,
-}
-
-#[tower_lsp::async_trait]
-impl LanguageServer for Backend {
-    async fn initialize(&self, params: InitializeParams) -> Result<InitializeResult> {
-        info!("TLSP initialize: {:?}", params.capabilities);
-
-        let code_editor = self.client.lock().await;
-        code_editor
-            .log_message(
-                MessageType::INFO,
-                "Aderyn LSP received an initialization request!",
-            )
-            .await;
-
-        Ok(InitializeResult {
-            capabilities: ServerCapabilities {
-                text_document_sync: Some(TextDocumentSyncCapability::Options(
-                    TextDocumentSyncOptions {
-                        open_close: Some(true),
-                        change: None,
-                        will_save: Some(false),
-                        will_save_wait_until: Some(false),
-                        save: None,
-                    },
-                )),
-                ..Default::default()
-            },
-            ..Default::default()
-        })
-    }
-
-    async fn initialized(&self, params: InitializedParams) {
-        info!("TLSP initialized: {:?}", params);
-
-        let code_editor = self.client.lock().await;
-        code_editor
-            .log_message(
-                MessageType::INFO,
-                "Aderyn LSP has been notified that the edtior's LSP client is initialized.",
-            )
-            .await;
-    }
-
-    async fn did_open(&self, params: DidOpenTextDocumentParams) {
-        info!("TLSP didOpen: {:?}", params);
-
-        let opened_file_uri = params.text_document.uri;
-
-        let code_editor = self.client.lock().await;
-        code_editor
-            .log_message(
-                MessageType::INFO,
-                format!(
-                    "Aderyn LSP has been notified that {} is opened",
-                    opened_file_uri
-                ),
-            )
-            .await;
-    }
-
-    async fn did_close(&self, params: DidCloseTextDocumentParams) {
-        info!("TLSP didClose: {:?}", params);
-
-        let opened_file_uri = params.text_document.uri;
-
-        let code_editor = self.client.lock().await;
-        code_editor
-            .log_message(
-                MessageType::INFO,
-                format!(
-                    "Aderyn LSP has been notified that {} is closed",
-                    opened_file_uri
-                ),
-            )
-            .await;
-    }
-
-    async fn shutdown(&self) -> Result<()> {
-        info!("TLSP shutdown");
-
-        let code_editor = self.client.lock().await;
-        code_editor
-            .log_message(MessageType::INFO, "Aderyn LSP has been shutdown")
-            .await;
-        Ok(())
-    }
-}
-
-fn spin_up_language_server(args: Args) {
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    rt.block_on(async {
-        let (tx, rx) = tokio::sync::mpsc::channel(1);
-
-        let config = Config::default()
-            .with_poll_interval(Duration::from_secs(2))
-            .with_compare_contents(false);
-
-        let mut watcher = RecommendedWatcher::new(
-            move |res| {
-                tx.blocking_send(res).unwrap();
-            },
-            config,
-        )
-        .unwrap();
-
-        watcher
-            .watch(
-                PathBuf::from(args.root.clone()).as_path(),
-                RecursiveMode::Recursive,
-            )
-            .unwrap();
-
-        //// Then run again only if file events are observed
-        //for result in rx {
-        //    match result {
-        //        Ok(_) => {
-        //
-        //            // do stuff here
-        //        }
-        //        Err(errors) => errors.iter().for_each(|error| println!("{error:?}")),
-        //    }
-        //    println!();
-        //}
-        _ = log_to_file(
-            "/Users/tilakmadichetti/Documents/OpenSource/my-first-vscode-lsp/lsp_server.log",
-            LevelFilter::Info,
-        );
-
-        let stdin = tokio::io::stdin();
-        let stdout = tokio::io::stdout();
-
-        let (service, socket) = LspService::new(move |client| {
-            let rx_arc = Arc::new(Mutex::new(rx));
-            let client_arc = Arc::new(Mutex::new(client));
-            let c2 = Arc::clone(&client_arc);
-            tokio::spawn(async move {
-                let t = Arc::clone(&rx_arc);
-                let c = Arc::clone(&client_arc);
-                let mut rxer = t.lock().await;
-                while let Some(change) = rxer.recv().await {
-                    if let Ok(_event) = change {
-                        // do something
-                        //
-
-                        let a = args.clone();
-                        let report_results = driver::drive_and_get_results(a);
-                        let report_results_lock = report_results.lock().await;
-
-                        let mut rcl = report_results_lock;
-                        // do some
-
-                        if let Some(_x) = &mut *rcl {
-                            // do some
-                        }
-                        let diagnostic = Diagnostic::new_simple(
-                            Range {
-                                start: Position {
-                                    line: 1,
-                                    character: 3,
-                                },
-                                end: Position {
-                                    line: 2,
-                                    character: 20,
-                                },
-                            },
-                            "BAD CODE".to_string(),
-                        );
-
-                        let c_lock = c.lock().await;
-
-                        c_lock
-                            .publish_diagnostics(
-                                Url::parse("file:///example/example.txt").unwrap(),
-                                vec![diagnostic],
-                                None,
-                            )
-                            .await;
-                    } else {
-                        warn!("Error from rexr receiver");
-                    }
-                }
-            });
-            Backend { client: c2 }
-        });
-
-        Server::new(stdin, stdout, socket).serve(service).await;
-    });
 }
